Restructure, add mock data, add StarRailService.test.ts (WIP)
