--- game/index.ts
@@ -1,3 +1,0 @@
-export * from "./Character";
-export * from "./Player";
-export * from "./Equipment";
