+++ src/api/StarRailApi.ts
