+++ src/api/StarRailApi.type.ts
