+++ src/api/index.ts
@@ -1,2 +1,2 @@
-export * from "./Api";
+export * from "./Api.ts";
 export * from "./StarRailApi.ts";
