+++ src/game/Character.ts
