+++ src/game/Equipment.ts
