+++ src/game/Mechanic.types.ts
