+++ src/game/index.ts
@@ -1,0 +1,4 @@
+export * from "./Character.ts";
+export * from "./Player.ts";
+export * from "./User.ts";
+export * from "./Equipment.ts";
