+++ tsconfig.json
@@ -18,5 +18,9 @@
     "types": [
       "bun-types" // add Bun global
     ]
-  }
+  },
+  "include": [
+    "./src/**/*.ts",
+    "src/_mocks/*.json"
+  ],
 }
